requirements.txt
google-cloud-aiplatform
google-generativeai
pydantic
python-dotenv
termcolor
faker
structlog
fastapi
uvicorn
websockets
